+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+Runtime changes:
+* Fix gray text output, using standard ANSI codes instead of a hardcoded RGB value
+
 ## v0.3.2
 
 API changes: